# cmd/server/docs/swagger.yaml
basePath: /
definitions:
...
  config.Config:
    type: object
...
  models.AudioRequest:
    properties:
      file:
        type: string
      model:
        type: string
      prompt:
        type: string
    type: object
  models.AudioResponse:
    properties:
      text:
        type: string
    type: object
  models.CacheControl:
    properties:
      type:
        description: e.g. "ephemeral"
        type: string
    type: object
  models.Choice:
    description: A single response choice.
    properties:
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/models.Message'
    type: object
  models.Embedding:
    properties:
      embedding:
        items:
          type: number
        type: array
      index:
        type: integer
      object:
        type: string
    type: object
  models.EmbeddingRequest:
    properties:
      input:
        type: string
      model:
        type: string
    type: object
  models.EmbeddingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Embedding'
        type: array
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  models.Image:
    properties:
      b64_json:
        type: string
      url:
        type: string
    type: object
  models.ImageRequest:
    properties:
      model:
        type: string
      "n":
        type: integer
      prompt:
        type: string
      size:
        type: string
    type: object
  models.ImageResponse:
    properties:
      created:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Image'
        type: array
    type: object
  models.JSONSchema:
    properties:
      name:
        type: string
      schema:
        additionalProperties: {}
        type: object
      strict:
        type: boolean
    type: object
  models.LLMRequest:
    description: Unified chat completion request compatible with OpenAI, Anthropic,
      and 100+ providers.
    properties:
      frequency_penalty:
        type: number
      max_tokens:
        description: Maximum tokens to generate.
        example: 4096
        type: integer
      messages:
        description: The conversation messages.
        items:
          $ref: '#/definitions/models.Message'
        type: array
      model:
        description: The model identifier (e.g. "gpt-4o", "claude-3-opus").
        example: gpt-4o
        type: string
      presence_penalty:
        type: number
      rag_enabled:
        type: boolean
      response_format:
        allOf:
        - $ref: '#/definitions/models.ResponseFormat'
        description: ResponseFormat controls structured output.
      stop:
        description: Stop sequences.
        items:
          type: string
        type: array
      stream:
        type: boolean
      temperature:
        description: Sampling temperature (0-2).
        example: 0.7
        type: number
      tools:
        items:
          $ref: '#/definitions/models.ToolDefinition'
        type: array
      top_p:
        description: Nucleus sampling (0-1).
        example: 1
        type: number
    type: object
  models.LLMResponse:
    description: Unified LLM response.
    properties:
      choices:
        items:
          $ref: '#/definitions/models.Choice'
        type: array
      created:
        example: 1700000000
        type: integer
      id:
        example: chatcmpl-123
        type: string
      model:
        example: gpt-4o
        type: string
      object:
        example: chat.completion
        type: string
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  models.Message:
    properties:
      cache_control:
        allOf:
        - $ref: '#/definitions/models.CacheControl'
        description: |-
          CacheControl enables prompt caching (Anthropic beta feature).
          When set to {"type":"ephemeral"}, Anthropic caches this message to reduce costs.
      content:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.MessageRole'
      tool_call_id:
        type: string
      tool_calls:
        items:
          $ref: '#/definitions/models.ToolCall'
        type: array
      tool_result:
        type: string
    type: object
  models.MessageRole:
    enum:
    - system
    - user
    - assistant
    - tool
    type: string
    x-enum-varnames:
    - RoleSystem
    - RoleUser
    - RoleAssistant
    - RoleTool
  models.ResponseFormat:
    properties:
      json_schema:
        $ref: '#/definitions/models.JSONSchema'
      type:
        description: '"json_schema" | "json_object" | "text"'
        type: string
    type: object
  models.ResponsesRequest:
    properties:
      input:
        type: string
      model:
        type: string
      previous_response_id:
        type: string
      tools:
        items:
          $ref: '#/definitions/models.ToolDefinition'
        type: array
    type: object
  models.ResponsesResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      output:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  models.ToolCall:
    properties:
      function:
        $ref: '#/definitions/models.ToolFunction'
      id:
        type: string
      type:
        type: string
    type: object
  models.ToolDefinition:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        additionalProperties: true
        type: object
    type: object
  models.ToolFunction:
    properties:
      arguments:
        type: string
      name:
        type: string
    type: object
  models.Usage:
    description: Token usage statistics for a completion request.
    properties:
      completion_tokens:
        example: 50
        type: integer
      prompt_tokens:
        example: 100
        type: integer
      total_tokens:
        example: 150
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: contact@ayoubzulfiqar.com
    name: AeroLLM Contributors
    url: https://github.com/ayoubzulfiqar/aerollm
  description: AeroLLM is a high-performance, multi-provider LLM gateway that serves
    as a drop-in replacement for the LiteLLM proxy. It supports 100+ LLM providers
    with advanced routing, virtual keys, semantic caching, spend analytics, and async
    observability callbacks.
  license:
    name: MIT
    url: https://github.com/ayoubzulfiqar/aerollm/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: AeroLLM API
  version: "1.0"
paths:
  /config/update:
    post:
      consumes:
      - application/json
      description: Accepts a partial or full config (JSON), validates it, and hot-reloads
        the provider registry without dropping active connections.
      parameters:
      - description: Updated configuration
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/config.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update config
      tags:
      - config
...
  /v1/audio/transcriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Audio transcription request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AudioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AudioResponse'
      summary: Transcribe audio
      tags:
      - audio
...
  /v1/embeddings:
    post:
      consumes:
      - application/json
      description: Generate embeddings for the given input.
      parameters:
      - description: Embeddings request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.EmbeddingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmbeddingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create embeddings
      tags:
      - embeddings
  /v1/images/generations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Image generation request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.ImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ImageResponse'
      summary: Generate image
      tags:
      - images
  /v1/messages:
    post:
      consumes:
      - application/json
      description: Anthropic-compatible /v1/messages endpoint (routes to chat completions).
      parameters:
      - description: Messages request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.LLMRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LLMResponse'
      summary: Anthropic-style messages
      tags:
      - chat
  /v1/responses:
    post:
      consumes:
      - application/json
      description: Create a response using the OpenAI Responses API.
      parameters:
      - description: Responses request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.ResponsesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponsesResponse'
      summary: Create response
      tags:
      - responses
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"